requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
isort==6.0.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
PyJWT==2.9.0
sqlparse==0.5.3
